remove unused functions and move super instructions to own file
